pyproject.toml
[tool.poetry]
name = "juniper"
version = "0.1.0"
description = ""
authors = ["Lawrence Lai <lawrence.c.lai@gmail.com>"]
readme = "README.md"
packages = [{include = "juniper"}]

[tool.poetry.dependencies]
python = "^3.10"
presidio-analyzer = {extras = ["stanza"], version = "^2.2.353"}
presidio-anonymizer = "^2.2.353"
streamlit = "^1.31.1"
streamlit-tags = "^1.2.8"
pandas = "^2.2.0"
python-dotenv = "^1.0.1"
st-annotated-text = "^4.0.1"
torch = "^2.2.0"
flair = "^0.13.1"
openai = "^1.12.0"
azure-ai-textanalytics = "^5.3.0"
transformers = "^4.37.2"
huggingface-hub = {extras = ["cli"], version = "^0.20.3"}
usaddress = "^0.5.10"
uszipcode = "^1.0.1"
nameparser = "^1.1.3"
python-levenshtein = "^0.25.0"
tqdm = "^4.66.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
